requirements.txt
streamlit
requests
numpy
Pillow
